Rejigged and added react checks for flow
